pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aps-acc-integration"
version = "0.1.0"
description = "Pull projects, issues, RFIs, and submittals from Autodesk Construction Cloud via APS REST APIs."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Ahmed Elfeki" }]
license = { text = "Proprietary" }
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "responses>=0.25",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-requests",
    "types-PyYAML",
]

[project.scripts]
aps-acc = "aps_acc.__main__:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
filterwarnings = ["error"]

[tool.coverage.run]
source = ["aps_acc"]
branch = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "SIM"]
ignore = ["E501"]  # handled by formatter

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
